Updated many things

+++ app/api/event/delete/[id]/route.ts
@@ -2,44 +2,59 @@ import { NextRequest, NextResponse } from "next/server";
 
 import EventModel from "@/models/Event";
 import { connectDB } from "@/lib/mongodb";
+import { withAuth } from "@/lib/withAuth";
 
 // ================= DELETE EVENT =================
 
-export async function DELETE(
-  req: NextRequest,
-  { params }: { params: Promise<{ id: string }> },
-) {
-  try {
-    await connectDB();
+// export async function DELETE(
+//   req: NextRequest,
+//   { params }: { params: Promise<{ id: string }> },
+// ) {
+export const DELETE = withAuth(
+  async (req, user, { params }: { params: Promise<{ id: string }> }) => {
+    try {
+      if (user.role !== "admin") {
+        return NextResponse.json(
+          {
+            success: false,
+            message: "Unauthorized",
+          },
+          {
+            status: 403,
+          },
+        );
+      }
+      await connectDB();
 
-    const { id } = await params;
+      const { id } = await params;
 
-    const deletedEvent = await EventModel.findByIdAndDelete(id);
+      const deletedEvent = await EventModel.findByIdAndDelete(id);
 
-    if (!deletedEvent) {
+      if (!deletedEvent) {
+        return NextResponse.json(
+          {
+            success: false,
+            message: "Event not found",
+          },
+          { status: 404 },
+        );
+      }
+
+      return NextResponse.json(
+        {
+          success: true,
+          message: "Event deleted successfully",
+        },
+        { status: 200 },
+      );
+    } catch (error: any) {
       return NextResponse.json(
         {
           success: false,
-          message: "Event not found",
+          message: error.message,
         },
-        { status: 404 },
+        { status: 500 },
       );
     }
-
-    return NextResponse.json(
-      {
-        success: true,
-        message: "Event deleted successfully",
-      },
-      { status: 200 },
-    );
-  } catch (error: any) {
-    return NextResponse.json(
-      {
-        success: false,
-        message: error.message,
-      },
-      { status: 500 },
-    );
-  }
-}
+  },
+);
